feat: allow the post and get to rpc discovery
