+++ ports/api/src/rpc/jsonrpc_endpoints.rs
@@ -1,11 +1,17 @@
 use super::handlers::admin_jsonrpc_post;
-use super::openrpc;
+use crate::rpc::openrpc::{generate_openrpc_spec, OpenRpcSpecConfig};
 
-use actix_web::web;
+use actix_web::{route, web, HttpResponse, Responder};
 
-/// Registers the admin JSON-RPC route and the OpenRPC spec route.
 pub fn configure(config: &mut web::ServiceConfig) {
-    config
-        .service(admin_jsonrpc_post)
-        .service(openrpc::openrpc_spec);
+    config.service(admin_jsonrpc_post).service(openrpc_spec);
+}
+
+#[route("", method = "GET", method = "POST")]
+pub async fn openrpc_spec(
+    config: web::Data<OpenRpcSpecConfig>,
+) -> impl Responder {
+    HttpResponse::Ok()
+        .content_type("application/json")
+        .json(generate_openrpc_spec(config.get_ref()))
 }
